Update session handling in the backend
Instead of using cookies and assigning user ID as
the value to the cookies, this commit utilizes
`express-session` for session management. The
cookie is set to expire after 30 seconds.

This eliminates the need to have a `sessionModel`
and `cookieController`.

One important thing to note is that the default
`store` is not compatible for production. We will
need to implement a different session store when
the app is deployed in the production environment
(https://github.com/expressjs/session#compatible-session-stores)